🚚 : 차트 hook, 배열 데이터 분리
