+++ src/features/realtime/components/chart/rec-line-chart/data.tsx
@@ -2,6 +2,12 @@ import { generateChartConfig } from "@/features/realtime/hooks/generate-chartCon
 import type { ChartConfig } from "@/shadcn/components/chart"
 import { chartData } from "./mock"
 
+export const recTimeRange = [
+	{ value: "30d", label: "Month" },
+	{ value: "90d", label: "Quarter" },
+	{ value: "180d", label: "Half" },
+]
+
 export function getChartConfig() {
 	const chartConfig: ChartConfig = generateChartConfig({ data: chartData })
