+++ src/features/realtime/components/chart/smp-line-chart/index.tsx
@@ -11,40 +11,16 @@ import {
 	ChartTooltipContent,
 } from "@/shadcn/components/chart"
 
-import { chartData, getChartConfig } from "./data"
+import { chartData, getChartConfig, smpTimeRange } from "./data"
 import { useState } from "react"
+import { dateFilteredData } from "@/features/realtime/hooks/date-range-filter"
 
 export function SmpLineChart() {
 	const chartConfig = getChartConfig()
 	const [timeRange, setTimeRange] = useState("1d")
 
-	const filteredData = chartData.filter((item) => {
-		const date = new Date(item.date)
-		const referenceDate = new Date(2024, 11, 28, 22, 0, 0)
-		const startDate = new Date(referenceDate)
-		const endDate = new Date(referenceDate)
-
-		let daysToSubtract = 90
-		if (timeRange === "1d") {
-			daysToSubtract = 1
-		} else if (timeRange === "7d") {
-			daysToSubtract = 7
-			endDate.setDate(startDate.getDate() - 2)
-		} else if (timeRange === "30d") {
-			daysToSubtract = 30
-			endDate.setDate(startDate.getDate() - 2)
-		}
-
-		startDate.setDate(endDate.getDate() - daysToSubtract)
-
-		return endDate >= date && date >= startDate
-	})
-
-	const timeRangeOptions = [
-		{ value: "1d", label: "Day" },
-		{ value: "7d", label: "Week" },
-		{ value: "30d", label: "Month" },
-	]
+	const filteredData = dateFilteredData({ chartData, timeRange, type: "smp" })
+	const timeRangeOptions = smpTimeRange
 
 	return (
 		<div>
