+++ src/features/realtime/components/types/smp-rec-linechart/index.ts
@@ -1,0 +1,5 @@
+type dateFilteredDataPT = {
+	chartData: { date: string; smp: number }[]
+	timeRange: string
+	type: string
+}
